+++ value.go
@@ -302,7 +302,7 @@ func (v *Value) Object() *Object {
 // print their definition.
 func (v *Value) String() string {
 	s := C.ValueToString(v.ptr)
-	defer C.free(unsafe.Pointer(s.data))
+	defer C.RtnStringRelease(s)
 	return C.GoStringN(s.data, C.int(s.length))
 }
 
@@ -678,4 +678,10 @@ func (v *Value) SharedArrayBufferGetContents() ([]byte, func(), error) {
 
 func (v *Value) StrictEquals(other *Value) bool {
 	return C.ValueStrictEquals(v.ptr, other.ptr) != 0
+}
+
+func (v *Value) TypeOf() string {
+	s := C.ValueTypeOf(v.ptr)
+	defer C.RtnStringRelease(s)
+	return C.GoStringN(s.data, C.int(s.length))
 }