+++ value.h
@@ -59,7 +59,9 @@ typedef struct {
 void ValueRelease(ValuePtr ptr);
 extern void* ValueToExternal(ValuePtr prt);
 extern uintptr_t ValueToExternalUintptr(ValuePtr prt);
+void RtnStringRelease(RtnString rtnString);
 extern RtnString ValueToString(ValuePtr ptr);
+extern RtnString ValueTypeOf(ValuePtr ptr);
 const uint32_t* ValueToArrayIndex(ValuePtr ptr);
 int ValueToBoolean(ValuePtr ptr);
 int32_t ValueToInt32(ValuePtr ptr);
@@ -171,5 +173,6 @@ extern BackingStorePtr SharedArrayBufferGetBackingStore(ValuePtr ptr);
 
 #ifdef __cplusplus
 }
+
 #endif
 #endif